+++ src/feature/SearchForm/api/__test__/useSearch.test.tsx
@@ -1,6 +1,6 @@
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
-
 import { renderHook, waitFor } from '@testing-library/react';
+
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 
 import useSearch from '../useSearch';
