+++ src/feature/SearchForm/components/SearchResultList/SearchResultItem.tsx
@@ -1,8 +1,7 @@
 import Avatar from '../../../../components/Avatar/Avatar';
 import Icons from '../../../../components/Icons';
 
-import { QueryDataProps } from '../../../../types/queryData';
-import { UserProps } from '../../../../types/user';
+import { QueryDataProps, SearchUserProps } from '../../../../types';
 
 import {
   StyledIconWrapper,
@@ -11,7 +10,7 @@ import {
 } from './SearchResultList.style';
 
 interface Props {
-  result: UserProps;
+  result: SearchUserProps;
   onSearchQuery: (queryData: QueryDataProps) => void;
 }
